# docs/phase-2/openapi.yaml
openapi: 3.1.0
info:
  title: PulseOS API
  version: "1.0.0"
  description: >
    The Real-Time Intelligence Operating System. Authoritative machine-readable
    contract. Conventions in 06-api-contracts.md. This is a specification excerpt
    covering the core MVP surface; full surface is generated per service.
  contact: { name: PulseOS Engineering, url: https://pulseos.com }
  license: { name: Proprietary }
servers:
  - url: https://api.pulseos.com/v1
  - url: https://staging-api.pulseos.com/v1
security:
  - bearerAuth: []
tags:
  - name: auth
  - name: dashboard
  - name: briefings
  - name: council
  - name: opportunities
  - name: memory
  - name: search
  - name: privacy

paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Password or OAuth login (issues access + refresh)
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TokenPair' } } } }
        '401': { $ref: '#/components/responses/Problem' }
        '429': { $ref: '#/components/responses/Problem' }
  /auth/refresh:
    post:
      tags: [auth]
      summary: Rotate refresh token, issue new access token
      security: []
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TokenPair' } } } }
        '401': { $ref: '#/components/responses/Problem' }
  /auth/mfa/verify:
    post:
      tags: [auth]
      summary: Verify a TOTP or WebAuthn challenge (step-up)
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/MfaVerify' } } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/TokenPair' } } } }
        '401': { $ref: '#/components/responses/Problem' }
  /auth/passkey/login/options:
    post:
      tags: [auth]
      summary: Begin WebAuthn assertion (returns challenge)
      security: []
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object } } } }

  /dashboard:
    get:
      tags: [dashboard]
      summary: Latest briefing + streams + analytics summary for the active user/tenant
      parameters:
        - { in: header, name: X-Tenant-Id, schema: { type: string, format: uuid }, required: false }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Dashboard' } } } }
        '401': { $ref: '#/components/responses/Problem' }

  /briefings:
    get:
      tags: [briefings]
      summary: List briefings (cursor paginated)
      parameters:
        - { in: query, name: cursor, schema: { type: string } }
        - { in: query, name: limit, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/BriefingPage' } } } }
  /briefings:generate:
    post:
      tags: [briefings]
      summary: Force-generate a fresh briefing (async)
      parameters:
        - { in: header, name: Idempotency-Key, schema: { type: string } }
      responses:
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } } }
  /briefings/{id}:
    get:
      tags: [briefings]
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/Briefing' } } } }
        '404': { $ref: '#/components/responses/Problem' }
  /briefing-items/{id}/feedback:
    post:
      tags: [briefings]
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/Feedback' } } }
      responses:
        '204': { description: Recorded }

  /council/reports/{id}:
    get:
      tags: [council]
      summary: Full auditable council report (summary, consensus, dissent, agent traces, evidence)
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/CouncilReport' } } } }
        '403': { $ref: '#/components/responses/Problem' }
  /council/analyze:
    post:
      tags: [council]
      summary: Ad-hoc council analysis of a subject (Pro+), async
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/AnalyzeRequest' } } }
      responses:
        '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } } }

  /opportunities:
    get:
      tags: [opportunities]
      parameters:
        - { in: query, name: domain, schema: { $ref: '#/components/schemas/OppDomain' } }
        - { in: query, name: status, schema: { type: string, enum: [open, acted, dismissed] } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/OpportunityPage' } } } }
  /opportunities/{id}:act:
    post:
      tags: [opportunities]
      summary: Mark opportunity acted-on (records North-Star WARU event)
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses:
        '204': { description: Recorded }

  /memory:
    get:
      tags: [memory]
      responses: { '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/MemoryItem' } } } } } }
    post:
      tags: [memory]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MemoryItemInput' } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/MemoryItem' } } } } }
  /memory/{id}:
    patch:
      tags: [memory]
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/MemoryItemInput' } } } }
      responses: { '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/MemoryItem' } } } } }
    delete:
      tags: [memory]
      summary: Delete memory item; propagates to embeddings + search
      parameters: [ { in: path, name: id, required: true, schema: { type: string, format: uuid } } ]
      responses: { '204': { description: Deleted } }

  /search:
    get:
      tags: [search]
      parameters:
        - { in: query, name: q, required: true, schema: { type: string } }
        - { in: query, name: types, schema: { type: string, description: "csv of entity types" } }
        - { in: query, name: cursor, schema: { type: string } }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/SearchResults' } } } }

  /privacy/export:
    post:
      tags: [privacy]
      summary: GDPR/CCPA data export (async); returns request id
      responses: { '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } } } }
  /privacy/delete:
    post:
      tags: [privacy]
      summary: GDPR erasure (async); crypto-shred + cascade
      responses: { '202': { description: Accepted, content: { application/json: { schema: { $ref: '#/components/schemas/Job' } } } } }

components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  responses:
    Problem:
      description: RFC 9457 problem
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        trace_id: { type: string }
        errors: { type: array, items: { type: object } }
    LoginRequest:
      type: object
      required: [method]
      properties:
        method: { type: string, enum: [password, oauth] }
        email: { type: string, format: email }
        password: { type: string, format: password }
        provider: { type: string }
        code: { type: string }
    TokenPair:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, default: Bearer }
        expires_in: { type: integer }
        mfa_required: { type: boolean }
    MfaVerify:
      type: object
      required: [factor, value]
      properties:
        factor: { type: string, enum: [totp, webauthn, recovery_code] }
        value: { type: string }
    Money:
      type: object
      properties: { amount: { type: number }, currency: { type: string, default: USD } }
    Confidence: { type: number, minimum: 0, maximum: 1 }
    OppDomain: { type: string, enum: [financial, career, athletic, education, business, consumer_savings] }
    Job:
      type: object
      properties:
        job_id: { type: string, format: uuid }
        status: { type: string, enum: [queued, running, done, failed] }
    Dashboard:
      type: object
      properties:
        briefing: { $ref: '#/components/schemas/Briefing' }
        streams: { type: array, items: { type: object } }
        analytics: { type: object }
    BriefingPage:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Briefing' } }
        next_cursor: { type: string, nullable: true }
    Briefing:
      type: object
      properties:
        id: { type: string, format: uuid }
        generated_at: { type: string, format: date-time }
        period: { type: string }
        summary: { type: string }
        items: { type: array, items: { $ref: '#/components/schemas/BriefingItem' } }
    BriefingItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        category: { type: string, enum: [action, opportunity, risk, time_saving, money_saving, event, forecast] }
        title: { type: string }
        rationale: { type: string }
        confidence: { $ref: '#/components/schemas/Confidence' }
        expected_value: { $ref: '#/components/schemas/Money' }
        ev_range: { type: object, properties: { low: { type: number }, high: { type: number } } }
        cost_of_inaction: { type: string }
        evidence: { type: array, items: { $ref: '#/components/schemas/Evidence' } }
        council_report_id: { type: string, format: uuid, nullable: true }
        rank: { type: integer }
    Evidence:
      type: object
      properties:
        signal_id: { type: string, format: uuid }
        source: { type: string }
        url: { type: string, nullable: true }
        snippet: { type: string }
        reliability: { $ref: '#/components/schemas/Confidence' }
    Feedback:
      type: object
      required: [verdict]
      properties:
        verdict: { type: string, enum: [useful, not_useful, wrong, acted_on] }
        note: { type: string }
    CouncilReport:
      type: object
      properties:
        id: { type: string, format: uuid }
        tier: { type: string, enum: [fast, full] }
        executive_summary: { type: string }
        consensus: { type: string }
        confidence: { $ref: '#/components/schemas/Confidence' }
        dissent: { type: array, items: { type: object, properties: { agent: { type: string }, position: { type: string }, rationale: { type: string } } } }
        evidence: { type: array, items: { $ref: '#/components/schemas/Evidence' } }
        agent_traces: { type: array, items: { $ref: '#/components/schemas/AgentTrace' } }
        recommended_actions: { type: array, items: { type: string } }
        estimated_impact: { $ref: '#/components/schemas/Money' }
    AgentTrace:
      type: object
      properties:
        agent: { type: string, enum: [economist, statistician, psychologist, domain_expert, contrarian, research_analyst, synthesizer] }
        stance: { type: string, enum: [support, oppose, neutral, abstain] }
        claims: { type: array, items: { type: object } }
        citations: { type: array, items: { type: object } }
    AnalyzeRequest:
      type: object
      required: [subject]
      properties:
        subject: { type: string }
        domain: { $ref: '#/components/schemas/OppDomain' }
        tier: { type: string, enum: [fast, full], default: full }
    OpportunityPage:
      type: object
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/Opportunity' } }
        next_cursor: { type: string, nullable: true }
    Opportunity:
      type: object
      properties:
        id: { type: string, format: uuid }
        domain: { $ref: '#/components/schemas/OppDomain' }
        title: { type: string }
        score: { $ref: '#/components/schemas/Confidence' }
        expected_value: { $ref: '#/components/schemas/Money' }
        confidence: { $ref: '#/components/schemas/Confidence' }
        risk_assessment: { type: string }
        recommended_action: { type: string }
        regulated: { type: boolean }
        status: { type: string, enum: [open, acted, dismissed] }
    MemoryItem:
      type: object
      properties:
        id: { type: string, format: uuid }
        kind: { type: string, enum: [interest, goal, profession, preference, episodic] }
        content: { type: string }
        confidence: { $ref: '#/components/schemas/Confidence' }
        source: { type: string }
        updated_at: { type: string, format: date-time }
    MemoryItemInput:
      type: object
      required: [kind, content]
      properties:
        kind: { type: string, enum: [interest, goal, profession, preference, episodic] }
        content: { type: string }
    SearchResults:
      type: object
      properties:
        results: { type: array, items: { type: object, properties: { type: { type: string }, id: { type: string }, title: { type: string }, snippet: { type: string }, score: { type: number } } } }
        next_cursor: { type: string, nullable: true }
        took_ms: { type: integer }